Move local completer into separate submodule
