+++ awscli/autocomplete/generator.py
@@ -11,7 +11,6 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 """Generates auto completion index."""
-from awscli.autocomplete import indexer
 
 
 class IndexGenerator(object):
