+++ awscli/autocomplete/local/indexer.py
