+++ scripts/gen-ac-index
@@ -4,7 +4,8 @@
 """
 import os
 
-from awscli.autocomplete import indexer
+from awscli.autocomplete.local import indexer
+from awscli.autocomplete import db
 from awscli.autocomplete.main import INDEX_DIR, INDEX_FILE
 from awscli import clidriver
 
@@ -16,7 +17,7 @@ def main():
     # only replace the entire file once we successfully regenerate the
     # index.
     temp_name = '%s.temp' % INDEX_FILE
-    db_connection = indexer.DatabaseConnection(temp_name)
+    db_connection = db.DatabaseConnection(temp_name)
     index = indexer.ModelIndexer(db_connection)
     driver = clidriver.create_clidriver()
     try:
