+++ tests/unit/autocomplete/__init__.py
@@ -10,7 +10,7 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-from awscli.autocomplete import model
+from awscli.autocomplete.local import model
 
 
 class InMemoryIndex(model.ModelIndex):
